change num_voices argument to voice_ids
