--- requirements.txt
@@ -1,6 +1,0 @@
-Keras==1.0.8
-music21==3.1.0
-numpy==1.11.0
-scipy==0.18.0
-tqdm==4.10.0
-h5py==2.6.0
